Optimised Matrix Rotation operation/Calculation/Transformation
Uses true FMA operations instead of individual scalar operations, which also compiles more cleanly into more concise and streamlined ASM as well.

zNear for the projection matrix had to be increased to compensate for odd clipping issues as a result of the new FMA Operations, however this could easily be due to the need to revise the winding order to Counter Clockwise instead of Clockwise.