Fix SwapChain fragment Glitches Due to incorrect modulus Operation
Due to an oversight with the modulus operation when resetting the swapchain count, if in exclusive fullscreen the framebuffer would exhibit seizure-inducing worthy hardware glitches due to displaying frames that were not fully rendered yet.

to clairify/elucidate; this was causes by an incorect increment of the swpachain counter.increment when comapired against the swapchain depth as a modulus, which if the swapchain value was not 2 or 4 would lead to the incorrect swapchain frame to be acquired, leading to severe graphical glitches.

This was fixed by incrementing the Swapchain counter after the Modulus Comparison/Operation and not before so the correct present order is restored and the framebuffer/SwapChain output/contents/raster/Rendered Output can be displayed/presented/rendered properly.