Big Cleanup: remove garbage Comments and code spam + removed Unused unused headers/and/or TUs + Code Smells
Some comments retained either due to reference purposes and/or due to some of the functions/Details being potentially used at a later date.

Removed VkTemp header as it seems to no longer be needed as VK_EXT_graphics_pipeline_library was added in alter version of the Vulkan headers.

Also removed unused Buffer/Descriptor set code due to being retain from an older version of the repo and will likely be faster to rewrite it properly with a much better code structure/layout instead of reusing it currently.